[task] Добавил проверку и сообщение.

+++ common/forms/Cat/CatLandConfigForm.php
@@ -59,17 +59,18 @@ class CatLandConfigForm extends Model
             "footer_tpl"     => $this->footer_tpl,
         ];
 
-        if($config_name != $this->config_name) {
+        if(!empty($config_name) && $config_name != $this->config_name) {
             unlink(\Yii::getAlias('@runtime') . '/cat/'.$config_name);
         }
 
         $config_file_extension = '.json';
         if (file_put_contents(Yii::getAlias('@frontend') . '/runtime/cat/' . $this->config_name . $config_file_extension,
             json_encode($json_config))) {
+            Yii::$app->session->setFlash('success', 'Удача! Данные сохранены');
             return true;
         }
 
-        return true;
+        return false;
     }
 
     public function storeOrUpdatePreviewConfig()
